+++ lib/cypher/order/index.ts
@@ -1,0 +1,3 @@
+import {OrderStatement} from "./OrderStatement";
+
+export const orderBy = OrderStatement.build;