+++ lib/model/Relationship.ts
@@ -178,7 +178,7 @@ export class Relationship<R extends RelationshipEntity, N extends NodeEntity<any
         baseQuery = appendReturn(baseQuery);
 
         if (this.orderStatement && !skipOrder) {
-            baseQuery = baseQuery.order(this.orderStatement);
+            baseQuery = baseQuery.pipe(this.orderStatement);
         }
 
         if (isPresent(this.skipCount) && !skipLimits) {